Documentation and multi-attachment support
Added help documentation and the changes from todays testing.  Originally I wasn't able to get the function to add multiple attachments, but reading ZD's API documentation all attachments submitted in an object have to have the same token.  This stores the first token and passes it in the URI for subsequent requests.